refactor: reorganize hybrid components into view/ subfolder mirroring native structure
Move Container and SizedBox from flat components/ into components/view/ to match
the native package layout. Add view/index.ts aggregating both exports, update tsup
entry points, package.json exports, and fix test import paths.
