+++ packages/hybrid/package.json
@@ -25,13 +25,9 @@
       "types": "./dist/core/index.d.ts",
       "import": "./dist/core/index.js"
     },
-    "./container": {
-      "types": "./dist/container/index.d.ts",
-      "import": "./dist/container/index.js"
-    },
-    "./sized-box": {
-      "types": "./dist/sized-box/index.d.ts",
-      "import": "./dist/sized-box/index.js"
+    "./view": {
+      "types": "./dist/view/index.d.ts",
+      "import": "./dist/view/index.js"
     }
   },
   "files": [
