+++ packages/hybrid/src/components/view/index.ts
@@ -1,0 +1,4 @@
+export { Container } from './container'
+export type { ContainerProps } from './container/container.type'
+export { SizedBox } from './sized-box'
+export type { SizedBoxProps } from './sized-box/sized-box.type'
