+++ packages/hybrid/src/components/view/sized-box/sized-box.tsx
