+++ packages/hybrid/src/components/view/sized-box/sized-box.type.ts
