--- packages/native/src/components/container/index.ts
@@ -1,2 +1,0 @@
-export { Container } from './container'
-export type { ContainerProps } from './container.type'
